(maint) Add integration tests for apply
This adds some basic integration tests for the apply statement. It moves
apply_spec to apply_compile spec since it confusingly mocks out actually
catalog compilation.
