+++ spec/fixtures/apply/basic/plans/failure.pp
@@ -1,0 +1,5 @@
+plan basic::failure (TargetSpec $nodes) {
+  return apply($nodes) {
+    package {'nonexistentpackagename': }
+  }
+}
